// src/modules/tasks/api/TasksApiRoutes.ts
export enum TasksApiRoutes {
  // tasks
  GET_TASK = '/api/crm/tasks/:id',
  GET_TASK_CARDS_BY_TIME = '/api/crm/tasks/by_time',
  GET_TASK_CARDS_BY_TIME_META = '/api/crm/tasks/by_time/meta',
  GET_TASK_CARDS_BY_TIME_FOR_CALENDAR = '/api/crm/tasks/by_time/calendar',
  GET_TASK_CARDS_BY_TIME_FOR_CALENDAR_META = '/api/crm/tasks/by_time/calendar/meta',
  GET_TASK_CARD_BY_TIME = '/api/crm/tasks/by_time/:type/:taskId',
  GET_TASK_CARDS_BOARD = '/api/crm/tasks/boards/:boardId',
  GET_TASK_CARDS_BOARD_META = '/api/crm/tasks/boards/:id/meta',
  GET_TASK_CARD = '/api/crm/tasks/cards/:taskId',
  GET_TASK_CARDS_LIST = '/api/crm/tasks/list/:boardId',
  GET_TASK_CARDS_LIST_META = '/api/crm/tasks/list/:boardId/meta',
  GET_TASK_CARDS_CALENDAR = '/api/crm/tasks/calendar/:boardId',
  GET_TASK_CARDS_CALENDAR_META = '/api/crm/tasks/calendar/:boardId/meta',
  ADD_TASK = '/api/crm/tasks',
  UPDATE_TASK = '/api/crm/tasks/:id',
  DELETE_TASK = '/api/crm/tasks/:id',
  // subtasks
  ADD_SUBTASK = '/api/crm/tasks/:id/subtasks',
  UPDATE_SUBTASK = '/api/crm/tasks/:taskId/subtasks/:subtaskId',
  DELETE_SUBTASK = '/api/crm/tasks/:taskId/subtasks/:subtaskId',
  // activities
  GET_ACTIVITY_CARDS_BY_TYPE = '/api/crm/activities/cards',
  GET_ACTIVITY_CARDS_META = '/api/crm/activities/cards/meta',
  GET_ACTIVITY_CARD = '/api/crm/activities/cards/:activityId',
  GET_ACTIVITIES = '/api/crm/activities',
  ADD_ACTIVITY = '/api/crm/activities',
  UPDATE_ACTIVITY = '/api/crm/activities/:id',
  DELETE_ACTIVITY = '/api/crm/activities/:id',
  GET_ACTIVITY_CARDS_FOR_CALENDAR = '/api/crm/activities/calendar',
  GET_ACTIVITY_CARDS_FOR_CALENDAR_META = '/api/crm/activities/calendar/meta',
  // task settings
  GET_TASK_SETTINGS_LIST = '/api/crm/task-settings',
  UPDATE_TASK_SETTINGS = '/api/crm/task-settings/:id',
  CREATE_TASK_SETTINGS = '/api/crm/task-settings',
  // task comments
  ADD_TASK_COMMENT = '/api/crm/tasks/:taskId/comments',
  GET_TASK_COMMENTS = '/api/crm/tasks/:taskId/comments',
  UPDATE_TASK_COMMENT = '/api/crm/tasks/:taskId/comments/:commentId',
  DELETE_TASK_COMMENT = '/api/crm/tasks/:taskId/comments/:commentId',
  LIKE_TASK_COMMENT = '/api/crm/tasks/:taskId/comments/:commentId/like',
  UNLIKE_TASK_COMMENT = '/api/crm/tasks/:taskId/comments/:commentId/unlike',
  // activity types
  GET_ACTIVITY_TYPES = '/api/crm/activity-types',
  ADD_ACTIVITY_TYPE = '/api/crm/activity-types',
  UPDATE_ACTIVITY_TYPE = '/api/crm/activity-types/:id',
  DELETE_ACTIVITY_TYPE = '/api/crm/activity-types/:id',
}